pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bt-module-programmer"
version = "0.1.0"
description = "PyQt6 desktop tool for configuring Bluetooth UART modules such as HC-05 and JDY-31."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "PyQt6>=6.6",
  "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.gui-scripts]
bt-module-programmer = "bt_programmer.main:run"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
